Reorganised folders and files names for Webhooks
